python/metrics-and-logging/prometheus-and-loggin-example/requirements.txt
cassandra-driver~=3.29.2